requirements.txt
cv2
numpy
torch>=1.0
torchvision
matplotlib
tensorflow
tensorboard
terminaltables
pillow==1.6
tqdm
opencv
shapely
colorama
visdom
yaml